let's color it

+++ src/algo_segment.c
@@ -6,7 +6,7 @@
 /*   By: floblanc <floblanc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/26 19:00:19 by maginist          #+#    #+#             */
-/*   Updated: 2019/10/03 15:56:41 by floblanc         ###   ########.fr       */
+/*   Updated: 2019/10/03 16:08:52 by floblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -137,8 +137,8 @@ void	select_seg_sens(t_map *map, int x, int y, int i)
 	z = -(i % map->max_x - map->max_x / 2) * sin(map->beta)
 		+ z * cos(map->beta);
 	z = map->x1;
-	map->x1 = (map->x1 * cos(map->phi) - map->y1 * sin(map->phi)) + map->move_x;
-	map->y1 = (z * sin(map->phi) + map->y1 * cos(map->phi)) + map->move_y;
+	map->x1 = (int)(map->x1 * cos(map->phi) - map->y1 * sin(map->phi)) + map->move_x;
+	map->y1 = (int)(z * sin(map->phi) + map->y1 * cos(map->phi)) + map->move_y;
 	if (x < map->x1 && y >= map->y1)
 		seg_rigth_top_writer(map, x, y, i);
 	else if (x <= map->x1 && y < map->y1)
